+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## 0.5.1 - Unreleased
 ### Fixes
 - Relative paths are now allowed in securitySchemes/OAuthFlow/tokenUrl (#130).
+- Schema validation errors will no longer print a stack trace (#131).
 
 
 ## 0.5.0 - 2020-08-05