+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## 0.5.1 - Unreleased
+## 0.5.1 - 2020-08-05
 ### Fixes
 - Relative paths are now allowed in securitySchemes/OAuthFlow/tokenUrl (#130).
 - Schema validation errors will no longer print a stack trace (#131).